[#1] First go at the RDF based DCAT harvester
Inspired by the existing ckanext-dcat harvester, but rather than relying
on parsing text files it uses rdflib to treat them as RDF graphs. It
uses the parser on parsers.py.

The main difference is that all CKAN datasets are generated on the
gather stage, as we have one single graph. The harvest objects store the
json dump of the ckan dicts, which are later created, udpated (or
deleted) in the import stage. This is done so all the reporting,
indexing, etc works as expected.