Reinstate --use-first and swap order
Without --use-first msgcat will merge the translations instead of
picking the first one. --use-first was removed in order to not
pick the translation of the older file (in case it needs to be
updated). So because we need --use-first we just swap the order of
the input files to get the overwrite we want.