Increment error count when gather error is not a json object
When the gather error is not a json object the error is more
fundamental, e.g. status code of the validation call is not 200 or
something like that. We should still treat that as an error
instead of just including the message and incrementing nothing.

The dcat_validation variable in the organization_list uses the
error count to set the boolean to true (no errors) or false
(errors). If there is a fundamental error, like a 400 error, the
validation should be marked as failed (false).